requirements.txt
pydantic>=2.6
typer[all]>=0.12
jinja2>=3.1
requests>=2.31
rich>=13.7
structlog>=23.2
fastapi>=0.110
uvicorn>=0.29
pyyaml>=6.0
